feat: Add majority element coding problem and test
This change adds a new coding problem to the `CodingProblems` directory. The new problem is to find the majority element in an array. A solution using the Boyer-Moore Voting Algorithm is provided in `CodingProblems/src/algorithms/majority_element.py`, and a corresponding test is included in `CodingProblems/tests/test_majority_element.py`. A `.gitignore` file was also added to the `CodingProblems` directory to prevent compiled Python files from being committed.